ios filters

+++ web/db/migrate/20180707195641_create_apartments.rb
@@ -4,7 +4,7 @@ class CreateApartments < ActiveRecord::Migration[5.1]
       t.string :name
       t.text :description
       t.decimal :floor_area_size
-      t.decimal :price_per_month
+      t.integer :price_per_month
       t.integer :number_of_rooms
       t.decimal :lat
       t.decimal :lon
